Add files via upload
In this notebook file I attempt to see how well I could predict the winner of March Madness games. There is a lot of data cleaning as there were about 10+ different csv's to be merged and formatted neatly for the model. I also went and found the 538 ratings which are basically just another experts rating for each team between 0- 100 (UConn would be something like 95.67 ). In summary, I used an XGBoost model to predict the winner of previous March Madness games based on basketball statistics, and the formerly mentioned expert ratings.